Crate test_raftstore

Modules

cluster

node

pd

router

server

transport_simulate

util

Structs

AddressMap

ChannelTransport

ChannelTransportCore

CloneFilterFactory

Cluster

CollectSnapshotFilter
CollectSnapshotFilter is a simulation transport filter to simulate the simultaneous delivery of multiple snapshots from different peers. It collects the snapshots from different peers and drop the subsequent snapshots from the same peers. Currently, if there are more than 1 snapshots in this filter, all the snapshots will be dilivered at once.

DefaultFilterFactory

DelayFilter

DropMessageFilter

DropPacketFilter

DropSnapshotFilter

IsolationFilterFactory

LeadingDuplicatedSnapshotFilter
Filter leading duplicated Snap.

LeadingFilter
Filters all filter_type packets until seeing the flush_type.

LeaseReadFilter

MessageTypeNotifier
Emits a notification for each given message type that it sees.

MockRaftStoreRouter

NodeCluster

PartitionFilterFactory

RandomLatencyFilter
RandomLatencyFilter is a transport filter to simulate randomized network latency. Based on a randomized rate, RandomLatencyFilter will decide whether to delay the sending of any message. It's could be used to simulate the message sending in a network with random latency, where messages could be delayed, disordered or lost.

RecvSnapshotFilter
Capture the first snapshot message.

RegionPacketFilter
Drop specified messages for the store with special region.

ServerCluster

SimulateTransport

SnapshotFilter

TestPdClient

Enums

Direction

Traits

Filter

FilterFactory

Simulator

Functions

async_read_index_on_peer

async_read_on_peer

batch_read_on_peer

bootstrap_with_first_region

check_lock_observer

check_messages

configure_for_disable_titan

configure_for_enable_titan

configure_for_encryption

configure_for_hibernate

configure_for_lease_read

configure_for_merge

configure_for_request_snapshot

configure_for_snapshot

configure_for_transfer_leader

create_test_engine

find_peer

ignore_merge_target_integrity

is_error_response

kv_pessimistic_lock

make_cb

make_cb_ext

must_check_lock_observer

must_check_txn_status

must_contains_error

must_error_read_on_peer

must_get

must_get_cf_equal

must_get_cf_none

must_get_equal

must_get_none

must_get_value

must_kv_commit

must_kv_pessimistic_lock

must_kv_pessimistic_rollback

must_kv_prewrite

must_new_and_configure_cluster

must_new_cluster

must_new_cluster_and_debug_client

must_new_cluster_and_kv_client

must_physical_scan_lock

must_read_on_peer

must_region_cleared

must_register_lock_observer

must_remove_lock_observer

new_admin_request

new_base_request

new_change_peer_request

new_change_peer_v2_request

new_compact_log_request

new_delete_cmd

new_delete_range_cmd

new_get_cf_cmd

new_get_cmd

new_incompatible_node_cluster

new_incompatible_server_cluster

new_learner_peer

new_mutation

new_node_cluster

new_pd_change_peer

new_pd_change_peer_v2

new_pd_merge_region

new_pd_transfer_leader

new_peer

new_prepare_merge

new_put_cf_cmd

new_put_cmd

new_read_index_cmd

new_region_detail_cmd

new_region_leader_cmd

new_request

new_server_cluster

new_snap_cmd

new_split_region

new_status_request

new_store

new_tikv_config

new_transfer_leader_cmd

put_cf_till_size

put_till_size
Keep putting random kvs until specified size limit is reached.

read_index_on_peer

read_on_peer

register_lock_observer

remove_lock_observer

sleep_ms

Type Definitions

SimulateEngine