Module test_raftstore::transport_simulate

Structs

CloneFilterFactory

CollectSnapshotFilter
CollectSnapshotFilter is a simulation transport filter to simulate the simultaneous delivery of multiple snapshots from different peers. It collects the snapshots from different peers and drop the subsequent snapshots from the same peers. Currently, if there are more than 1 snapshots in this filter, all the snapshots will be dilivered at once.

DefaultFilterFactory

DelayFilter

DropMessageFilter

DropPacketFilter

DropSnapshotFilter

IsolationFilterFactory

LeadingDuplicatedSnapshotFilter
Filter leading duplicated Snap.

LeadingFilter
Filters all filter_type packets until seeing the flush_type.

LeaseReadFilter

MessageTypeNotifier
Emits a notification for each given message type that it sees.

PartitionFilter

PartitionFilterFactory

RandomLatencyFilter
RandomLatencyFilter is a transport filter to simulate randomized network latency. Based on a randomized rate, RandomLatencyFilter will decide whether to delay the sending of any message. It's could be used to simulate the message sending in a network with random latency, where messages could be delayed, disordered or lost.

RecvSnapshotFilter
Capture the first snapshot message.

RegionPacketFilter
Drop specified messages for the store with special region.

SimulateTransport

SnapshotFilter

Enums

Direction

Traits

Filter

FilterFactory

Functions

check_messages

filter_send