Trait test_raftstore::transport_simulate::Filter

pub trait Filter: Send + Sync {
    fn before(&self, msgs: &mut Vec<RaftMessage>) -> Result<()>;

    fn after(&self, res: Result<()>) -> Result<()> { ... }
}

Required methods

fn before(&self, msgs: &mut Vec<RaftMessage>) -> Result<()>

before is run before sending the messages.

Loading content...
Provided methods

fn after(&self, res: Result<()>) -> Result<()>

after is run after sending the messages, so that the returned value could be changed if necessary.

Loading content...